# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Job Automation API
  description: REST API for job search, application tracking, and auto-apply automation
  version: 1.4.0
  contact:
    email: qws941@kakao.com
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://resume.jclee.me/job
    description: Production (Cloudflare Worker)
...
security: []
paths:
...
  /api/applications:
    get:
      tags: [Applications]
      summary: List applications
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              [
                pending,
                saved,
                applied,
                viewed,
                in_progress,
                interview,
                offer,
                rejected,
                withdrawn,
                expired,
              ]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Application list
          content:
            application/json:
              schema:
                type: object
                properties:
                  applications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Application'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

    post:
      tags: [Applications]
      summary: Create application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                job:
                  $ref: '#/components/schemas/Job'
                options:
                  type: object
      responses:
        '201':
          description: Application created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
  /api/applications/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    get:
      tags: [Applications]
      summary: Get application by ID
      responses:
        '200':
          description: Application details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '404':
          description: Not found

    put:
      tags: [Applications]
      summary: Update application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  application:
                    $ref: '#/components/schemas/Application'

    delete:
      tags: [Applications]
      summary: Delete application
      responses:
        '200':
          description: Deleted
  /api/applications/{id}/status:
    put:
      tags: [Applications]
      summary: Update application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    [
                      pending,
                      saved,
                      applied,
                      viewed,
                      in_progress,
                      interview,
                      offer,
                      rejected,
                      withdrawn,
                      expired,
                    ]
                note:
                  type: string
                notifyAutomation:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Status updated
...
  /api/auth/status:
    get:
      tags: [Auth]
      summary: Current admin session status
      operationId: getAuthStatus
      responses:
        '200':
          description: Session status
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
                  expiresAt:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Clear admin session cookie
      operationId: logoutAuth
      responses:
        '200':
          description: Session cleared
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/auth/set:
    post:
      tags: [Auth]
      summary: Manually set platform auth (admin)
      operationId: setAuth
      responses:
        '200':
          description: Auth updated
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/auth/{platform}:
    delete:
      tags: [Auth]
      summary: Clear platform-specific session
      operationId: deleteAuthPlatform
      parameters:
        - in: path
          name: platform
          schema: { type: string }
          required: true
      responses:
        '200':
          description: Platform session cleared
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Workflows (apps/job-dashboard/src/routes/workflows.js)
...
  /api/workflows/application:
    post:
      tags: [Workflows]
      summary: Trigger application Cloudflare Workflow
      operationId: triggerApplicationWorkflow
      responses:
        '202':
          description: Workflow started
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    Job:
      type: object
      properties:
        id:
          oneOf:
            - type: integer
            - type: string
        company:
          type: string
        position:
          type: string
        source:
          $ref: '#/components/schemas/AutoApplyPlatform'
        matchScore:
          type: integer
        url:
          type: string
    Application:
      type: object
      properties:
        id:
          type: string
        job:
          $ref: '#/components/schemas/Job'
        status:
          type: string
          enum:
            [
              pending,
              saved,
              applied,
              viewed,
              in_progress,
              interview,
              offer,
              rejected,
              withdrawn,
              expired,
            ]
        appliedAt:
          type: string
          format: date-time
        notes:
          type: array
          items:
            type: string
        source:
          type: string
    ApplicationUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: array
          items:
            type: string
...
    AutoApplyPlatform:
      type: string
      enum: [wanted, linkedin, remember, jobkorea, saramin, greenhouse, lever, ashby]
...